# target/classes/openapi/api-out.yaml
openapi: 3.0.1
info:
  title: ValkyrAI CORE API
  description: The API for accessing ValkyrAI services
  termsOfService: https://valkyrlabs.com/tos/
  contact: {email: info@valkyrlabs.com}
  license: {name: AGPL 3.0, url: 'https://www.gnu.org/licenses/agpl-3.0.html'}
  version: 0.9.25
...
servers:
- {url: 'http://localhost:8080/v1', description: local server (uses live data)}
- {url: 'https://valkyrlabs.com/v1', description: production server}
...
paths:
...
  /Address:
    get:
      tags: [Address]
      summary: Retrieve a list of Addresss
      description: Retrieves a list of Addresss.
      x-spring-cacheable: true
      operationId: getAddressList
      responses:
        '200':
          description: Successfully retrieved the list of Addresss.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Address'}
        '404': {description: No Addresss were found.}
        '500': {description: Internal server error.}
    post:
      tags: [Address]
      summary: Create a new Address
      description: Creates a new Address.
      operationId: postAddress
      requestBody:
        description: Address details.
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Address'}
      responses:
        '201':
          description: Address created successfully.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Address'}
        '400': {description: Invalid input.}
        '404': {description: Address not found.}
        '500': {description: Internal server error.}
  /Address/{id}:
    get:
      tags: [Address]
      summary: Retrieve a single Address
      description: Retrieves a single Address for a specific uid.
      x-spring-cacheable: true
      operationId: getAddress
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the Address.
      responses:
        '200':
          description: Successfully retrieved the Address.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Address'}
        '404': {description: Address not found.}
        '500': {description: Internal server error.}
    put:
      tags: [Address]
      summary: Update an existing Address
      description: Updates an existing Address.
      operationId: updateAddress
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the Address.
      requestBody:
        description: Updated Address details.
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Address'}
      responses:
        '200':
          description: Address updated successfully.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Address'}
        '400': {description: Invalid input.}
        '404': {description: Address not found.}
        '500': {description: Internal server error.}
    delete:
      tags: [Address]
      summary: Delete a Address.
      description: Deletes a specific Address.
      operationId: deleteAddress
      parameters:
      - name: id
        schema: {type: string, format: uuid}
        in: path
        required: true
        description: Unique identifier for the Address.
      responses:
        '204': {description: Address deleted successfully.}
        '404': {description: Address not found.}
        '500': {description: Internal server error.}
...
components:
...
  schemas:
...
    Address:
      required: [city, name, state, street1, postal_code]
      type: object
      description: TODO Address CLASS DESCRIPTION
      properties:
        hasWifi: {type: boolean}
        principalId: {type: string, format: uuid}
        addressType:
          type: string
          description: The type of address
          enum: [home, mail, work, school, other]
          example: home
        name:
          type: string
          description: full name of addressee
          example: Henry Johnson
          x-thorapi-dataField: advanced=true
          x-thorapi-secureField: true
          x-field-extra-annotation: |-
            @DataField(unique = false, hidden = false, advanced = false, relationshipTable="advanced=true")
            @SecureField(encryptionType = SecureField.EncryptionType.SYMMETRIC, strength = 5)
        street1: {type: string, description: street address line 1, example: 123 Daydream
            Blvd.}
        street2: {type: string, description: street address line 2, example: Apt.
            7}
        city:
          type: string
          description: full name of city
          example: San Francisco
          x-thorapi-dataField: advanced=true
          x-thorapi-secureField: true
          x-field-extra-annotation: |-
            @DataField(unique = false, hidden = false, advanced = false, relationshipTable="advanced=true")
            @SecureField(encryptionType = SecureField.EncryptionType.SYMMETRIC, strength = 5)
        state:
          minLength: 2
          type: string
          description: State of residence 2-character code (USA)
          example: CA
          x-thorapi-dataField: advanced=true
          x-thorapi-secureField: true
          x-field-extra-annotation: |-
            @DataField(unique = false, hidden = false, advanced = false, relationshipTable="advanced=true")
            @SecureField(encryptionType = SecureField.EncryptionType.SYMMETRIC, strength = 5)
        country: {maxLength: 3, minLength: 3, example: USA, type: string, description: 3
            character country code (upper case), x-field-extra-annotation: '@Column(length
            = 3)'}
        postal_code: {maxLength: 10, minLength: 10, type: string, pattern: '^\d{5}-\d{4}$',
          description: 5+4 digit zip code, example: 94210-0000, x-field-extra-annotation: '@Column(length
            = 10)'}
        id:
          description: Unique identifier for object in the system
          type: string
          x-field-extra-annotation: "@Id  \n        @GeneratedValue(generator = \"\
            UUID\")\n"
          format: uuid
          example: 3773d666-dea1-4d02-b430-11ca52e266a8
        ownerId: {description: UUID of owner of the object in the system, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.CREATED_BY,
            enabled = true)', format: uuid, example: 5297b19d-fd2a-4007-9aef-f0b7264cd539}
        createdDate: {description: Date of object creation, type: string, x-field-extra-annotation: '        @AuditingField(fieldType
            = AuditingField.FieldType.CREATED_DATE, enabled = true)', format: date-time,
          example: '2025-01-03T16:21:41.594-0800'}
        keyHash: {description: 'Data, including hash of the key(s) used to encrypt
            this record.', type: string}
        lastAccessedById: {description: Last user to access object, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.LAST_ACCESSED_BY,
            enabled = true)', format: uuid, example: ef297a29-e742-4535-82c6-6ef6211bca12}
        lastAccessedDate: {description: Timestamp of last access of object, type: string,
          x-field-extra-annotation: '        @AuditingField(fieldType = AuditingField.FieldType.LAST_ACCESSED_DATE,
            enabled = true)', format: date-time, example: '2025-01-03T16:21:41.595-0800'}
        lastModifiedById: {description: Unique identifier for user who last modifed
            the object in the system, type: string, x-field-extra-annotation: '        @AuditingField(fieldType
            = AuditingField.FieldType.LAST_MODIFIED_BY, enabled = true)', format: uuid,
          example: ba749007-a05a-4505-a434-d92c7cf9712d}
        lastModifiedDate: {description: Date of last object modification, type: string,
          x-field-extra-annotation: '    @AuditingField(fieldType = AuditingField.FieldType.LAST_MODIFIED_DATE,
            enabled = true)', format: date-time, example: '2025-01-03T16:21:41.595-0800'}
      xml: {name: Address}
...
  securitySchemes:
    thorapi_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://{username}.valkyrlabs.com:{port}/oauth
          scopes: {'write:items': modify owned items, 'read:items': read owned items}